Data.String.Here.Interpolated

Description

Interpolated here docs

Documentation

i :: QuasiQuoter

Quote a here doc with embedded antiquoted expressions

Any expression occurring between ${ and } (for which the type must have Show and Typeable instances) will be interpolated into the quoted string.

Characters preceded by a backslash are treated literally. This enables the inclusion of the literal substring ${ within your quoted text by writing it as \${. The literal sequence \${ may be written as \\${.

iTrim :: QuasiQuoter

Like i, but with leading and trailing whitespace trimmed

template :: QuasiQuoter

Quote the contents of a file as with i

This enables usage as a simple template engine